+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 
+## [2022-08-02] v0.2.1
+
+fix(🐞): Windows slash path
+
+- last commit 7f53465
+- PR: #6
+
+
+## [2022-08-02] v0.2.0
+
+- d81dc3a test: v0.2.0
+- 1c78b01 feat: only import asset file default module
+- 7fd0c05 refactor: better logic
+
 ## [2022-06-27] v0.1.3
 
 - 898c179 docs: v0.1.3